core: Huge mod reorg
