+++ psst-cli/src/main.rs
@@ -1,12 +1,11 @@
 use psst_core::{
-    audio_normalize::NormalizationLevel,
-    audio_output::AudioOutput,
-    audio_player::{PlaybackConfig, PlaybackItem, Player, PlayerCommand, PlayerEvent},
+    audio::{normalize::NormalizationLevel, output::AudioOutput},
     cache::{Cache, CacheHandle},
     cdn::{Cdn, CdnHandle},
     connection::Credentials,
     error::Error,
     item_id::{ItemId, ItemIdType},
+    player::{item::PlaybackItem, PlaybackConfig, Player, PlayerCommand, PlayerEvent},
     session::{SessionConfig, SessionService},
 };
 use std::{env, io, io::BufRead, path::PathBuf, thread};
@@ -57,7 +56,7 @@ fn play_item(
     let mut player = Player::new(session, cdn, cache, config, &output);
 
     let _ui_thread = thread::spawn({
-        let player_sender = player.event_sender();
+        let player_sender = player.sender();
 
         player_sender
             .send(PlayerEvent::Command(PlayerCommand::LoadQueue {
@@ -100,7 +99,7 @@ fn play_item(
         }
     });
 
-    for event in player.event_receiver() {
+    for event in player.receiver() {
         player.handle(event);
     }
     output.sink().close();
