+++ psst-core/src/audio/mod.rs
@@ -1,0 +1,5 @@
+pub mod decode;
+pub mod decrypt;
+pub mod normalize;
+pub mod output;
+pub mod resample;
