+++ psst-core/src/lib.rs
@@ -1,25 +1,15 @@
 #![allow(clippy::new_without_default)]
 
-pub mod access_token;
 pub mod actor;
-pub mod audio_decode;
-pub mod audio_decrypt;
-pub mod audio_file;
-pub mod audio_key;
-pub mod audio_normalize;
-pub mod audio_output;
-pub mod audio_player;
-pub mod audio_queue;
-pub mod audio_resample;
+pub mod audio;
 pub mod cache;
 pub mod cdn;
 pub mod connection;
 pub mod error;
 pub mod item_id;
-pub mod mercury;
 pub mod metadata;
+pub mod player;
 pub mod session;
-pub mod stream_storage;
 pub mod util;
 
 pub use psst_protocol as protocol;
