+++ psst-core/src/player/mod.rs
@@ -1,3 +1,9 @@
+pub mod file;
+pub mod item;
+pub mod queue;
+mod storage;
+mod worker;
+
 use std::{
     mem,
     sync::{
@@ -10,24 +16,21 @@ use std::{
 };
 
 use crossbeam_channel::{unbounded, Receiver, Sender};
-use symphonia::core::audio::{SampleBuffer, SignalSpec};
 
 use crate::{
-    actor::{Actor, ActorHandle, ActorOp},
-    audio_decode::AudioDecoder,
-    audio_file::{AudioFile, AudioPath},
-    audio_key::AudioKey,
-    audio_normalize::NormalizationLevel,
-    audio_output::{AudioOutput, AudioSink},
-    audio_queue::{Queue, QueueBehavior},
-    audio_resample::{AudioResampler, ResamplingAlgo, ResamplingSpec},
+    actor::Actor,
+    audio::output::{AudioOutput, AudioSink},
     cache::CacheHandle,
     cdn::CdnHandle,
     error::Error,
-    item_id::{ItemId, ItemIdType},
-    metadata::{Fetch, ToAudioPath},
-    protocol::metadata::Track,
     session::SessionService,
+};
+
+use self::{
+    file::AudioPath,
+    item::{LoadedPlaybackItem, PlaybackItem},
+    queue::{Queue, QueueBehavior},
+    worker::{Decode, Decoding, DecodingWorker},
 };
 
 const PREVIOUS_TRACK_THRESHOLD: Duration = Duration::from_secs(3);
@@ -48,137 +51,6 @@ impl Default for PlaybackConfig {
     }
 }
 
-#[derive(Debug, Clone, Copy, Eq, PartialEq)]
-pub struct PlaybackItem {
-    pub item_id: ItemId,
-    pub norm_level: NormalizationLevel,
-}
-
-impl PlaybackItem {
-    fn load(
-        &self,
-        session: &SessionService,
-        cdn: CdnHandle,
-        cache: CacheHandle,
-        config: &PlaybackConfig,
-    ) -> Result<LoadedPlaybackItem, Error> {
-        let path = load_audio_path(self.item_id, session, &cache, config)?;
-        let key = load_audio_key(&path, session, &cache)?;
-        let file = AudioFile::open(path, cdn, cache)?;
-        let (source, norm_data) = file.audio_source(key)?;
-        let norm_factor = norm_data.factor_for_level(self.norm_level, config.pregain);
-        Ok(LoadedPlaybackItem {
-            file,
-            source,
-            norm_factor,
-        })
-    }
-}
-
-fn load_audio_path(
-    item_id: ItemId,
-    session: &SessionService,
-    cache: &CacheHandle,
-    config: &PlaybackConfig,
-) -> Result<AudioPath, Error> {
-    match item_id.id_type {
-        ItemIdType::Track => {
-            load_audio_path_from_track_or_alternative(item_id, session, cache, config)
-        }
-        ItemIdType::Podcast | ItemIdType::Unknown => unimplemented!(),
-    }
-}
-
-fn load_audio_path_from_track_or_alternative(
-    item_id: ItemId,
-    session: &SessionService,
-    cache: &CacheHandle,
-    config: &PlaybackConfig,
-) -> Result<AudioPath, Error> {
-    let track = load_track(item_id, session, cache)?;
-    let country = get_country_code(session, cache);
-    let path = match country {
-        Some(user_country) if track.is_restricted_in_region(&user_country) => {
-            // The track is regionally restricted and is unavailable.  Let's try to find an
-            // alternative track.
-            let alt_id = track
-                .find_allowed_alternative(&user_country)
-                .ok_or(Error::AudioFileNotFound)?;
-            let alt_track = load_track(alt_id, session, cache)?;
-            let alt_path = alt_track
-                .to_audio_path(config.bitrate)
-                .ok_or(Error::AudioFileNotFound)?;
-            // We've found an alternative track with a fitting audio file.  Let's cheat a
-            // little and pretend we've obtained it from the requested track.
-            // TODO: We should be honest and display the real track information.
-            AudioPath {
-                item_id,
-                ..alt_path
-            }
-        }
-        _ => {
-            // Either we do not have a country code loaded or the track is available, return
-            // it.
-            track
-                .to_audio_path(config.bitrate)
-                .ok_or(Error::AudioFileNotFound)?
-        }
-    };
-    Ok(path)
-}
-
-fn get_country_code(session: &SessionService, cache: &CacheHandle) -> Option<String> {
-    if let Some(cached_country_code) = cache.get_country_code() {
-        Some(cached_country_code)
-    } else {
-        let country_code = session.connected().ok()?.get_country_code()?;
-        if let Err(err) = cache.save_country_code(&country_code) {
-            log::warn!("failed to save country code to cache: {:?}", err);
-        }
-        Some(country_code)
-    }
-}
-
-fn load_track(
-    item_id: ItemId,
-    session: &SessionService,
-    cache: &CacheHandle,
-) -> Result<Track, Error> {
-    if let Some(cached_track) = cache.get_track(item_id) {
-        Ok(cached_track)
-    } else {
-        let track = Track::fetch(session, item_id)?;
-        if let Err(err) = cache.save_track(item_id, &track) {
-            log::warn!("failed to save track to cache: {:?}", err);
-        }
-        Ok(track)
-    }
-}
-
-fn load_audio_key(
-    path: &AudioPath,
-    session: &SessionService,
-    cache: &CacheHandle,
-) -> Result<AudioKey, Error> {
-    if let Some(cached_key) = cache.get_audio_key(path.item_id, path.file_id) {
-        Ok(cached_key)
-    } else {
-        let key = session
-            .connected()?
-            .get_audio_key(path.item_id, path.file_id)?;
-        if let Err(err) = cache.save_audio_key(path.item_id, path.file_id, &key) {
-            log::warn!("failed to save audio key to cache: {:?}", err);
-        }
-        Ok(key)
-    }
-}
-
-pub struct LoadedPlaybackItem {
-    file: AudioFile,
-    source: AudioDecoder,
-    norm_factor: f32,
-}
-
 pub struct Player {
     state: PlayerState,
     preload: PreloadState,
@@ -187,8 +59,8 @@ pub struct Player {
     cache: CacheHandle,
     config: PlaybackConfig,
     queue: Queue,
-    event_sender: Sender<PlayerEvent>,
-    event_receiver: Receiver<PlayerEvent>,
+    sender: Sender<PlayerEvent>,
+    receiver: Receiver<PlayerEvent>,
     audio_output_sink: AudioSink<f32>,
     audio_volume: VolumeLevel,
     consecutive_loading_failures: usize,
@@ -202,14 +74,14 @@ impl Player {
         config: PlaybackConfig,
         audio_output: &AudioOutput,
     ) -> Self {
-        let (event_sender, event_receiver) = unbounded();
+        let (sender, receiver) = unbounded();
         Self {
             session,
             cdn,
             cache,
             config,
-            event_sender,
-            event_receiver,
+            sender,
+            receiver,
             audio_output_sink: audio_output.sink(),
             audio_volume: VolumeLevel::new(),
             state: PlayerState::Stopped,
@@ -219,12 +91,12 @@ impl Player {
         }
     }
 
-    pub fn event_sender(&self) -> Sender<PlayerEvent> {
-        self.event_sender.clone()
+    pub fn sender(&self) -> Sender<PlayerEvent> {
+        self.sender.clone()
     }
 
-    pub fn event_receiver(&self) -> Receiver<PlayerEvent> {
-        self.event_receiver.clone()
+    pub fn receiver(&self) -> Receiver<PlayerEvent> {
+        self.receiver.clone()
     }
 
     pub fn handle(&mut self, event: PlayerEvent) {
@@ -325,7 +197,7 @@ impl Player {
                 *position = new_position;
             }
             _ => {
-                log::warn!("received unexpected position report");
+                log::warn!("received ununwraped position report");
             }
         }
         const PRELOAD_BEFORE_END_OF_TRACK: Duration = Duration::from_secs(30);
@@ -375,7 +247,7 @@ impl Player {
         }
         // Item is not preloaded yet, load it in a background thread.
         let loading_handle = thread::spawn({
-            let event_sender = self.event_sender.clone();
+            let event_sender = self.sender.clone();
             let session = self.session.clone();
             let cdn = self.cdn.clone();
             let cache = self.cache.clone();
@@ -384,12 +256,10 @@ impl Player {
                 let result = item.load(&session, cdn, cache, &config);
                 event_sender
                     .send(PlayerEvent::Loaded { item, result })
-                    .expect("Failed to send PlayerEvent::Loaded");
+                    .unwrap();
             }
         });
-        self.event_sender
-            .send(PlayerEvent::Loading { item })
-            .expect("Failed to send PlayerEvent::Loading");
+        self.sender.send(PlayerEvent::Loading { item }).unwrap();
         self.state = PlayerState::Loading {
             item,
             _loading_handle: loading_handle,
@@ -401,7 +271,7 @@ impl Player {
             return;
         }
         let loading_handle = thread::spawn({
-            let event_sender = self.event_sender.clone();
+            let event_sender = self.sender.clone();
             let session = self.session.clone();
             let cdn = self.cdn.clone();
             let cache = self.cache.clone();
@@ -410,7 +280,7 @@ impl Player {
                 let result = item.load(&session, cdn, cache, &config);
                 event_sender
                     .send(PlayerEvent::Preloaded { item, result })
-                    .expect("Failed to send PlayerEvent::Preloaded");
+                    .unwrap();
             }
         });
         self.preload = PreloadState::Preloading {
@@ -427,23 +297,21 @@ impl Player {
         log::info!("starting playback");
         let path = loaded_item.file.path();
         let position = Duration::default();
-        let worker = PlaybackWorker {
-            actor: Decoding::spawn_default({
-                let events = self.event_sender.clone();
-                let sink = self.audio_output_sink.clone();
-                let volume = self.audio_volume.clone();
-                move |this| Decoding::new(loaded_item, events, this, sink, volume)
-            }),
-        };
-        worker.start();
+        let worker = Decoding::spawn_default({
+            let events = self.sender.clone();
+            let sink = self.audio_output_sink.clone();
+            let volume = self.audio_volume.clone();
+            move |this| Decoding::new(loaded_item, events, this, sink, volume)
+        });
+        worker.send(Decode::Start).unwrap();
         self.state = PlayerState::Playing {
             path,
             position,
-            worker,
+            worker: DecodingWorker { actor: worker },
         };
-        self.event_sender
+        self.sender
             .send(PlayerEvent::Playing { path, position })
-            .expect("Failed to send PlayerEvent::Playing");
+            .unwrap();
     }
 
     fn pause(&mut self) {
@@ -459,10 +327,10 @@ impl Player {
                 worker,
             } => {
                 log::info!("pausing playback");
-                worker.stop();
-                self.event_sender
+                worker.actor.send(Decode::Stop).unwrap();
+                self.sender
                     .send(PlayerEvent::Pausing { path, position })
-                    .expect("Failed to send PlayerEvent::Paused");
+                    .unwrap();
                 self.state = PlayerState::Paused {
                     path,
                     position,
@@ -488,10 +356,10 @@ impl Player {
                 worker,
             } => {
                 log::info!("resuming playback");
-                worker.start();
-                self.event_sender
+                worker.actor.send(Decode::Start).unwrap();
+                self.sender
                     .send(PlayerEvent::Resuming { path, position })
-                    .expect("Failed to send PlayerEvent::Resuming");
+                    .unwrap();
                 self.state = PlayerState::Playing {
                     path,
                     position,
@@ -537,9 +405,7 @@ impl Player {
     }
 
     fn stop(&mut self) {
-        self.event_sender
-            .send(PlayerEvent::Stopped)
-            .expect("Failed to send PlayerEvent::Stopped");
+        self.sender.send(PlayerEvent::Stopped).unwrap();
         self.state = PlayerState::Stopped;
         self.queue.clear();
         self.consecutive_loading_failures = 0;
@@ -549,7 +415,7 @@ impl Player {
         if let PlayerState::Playing { worker, .. } | PlayerState::Paused { worker, .. } =
             &mut self.state
         {
-            worker.seek(position);
+            worker.actor.send(Decode::Seek(position)).unwrap();
         }
     }
 
@@ -665,12 +531,12 @@ enum PlayerState {
     Playing {
         path: AudioPath,
         position: Duration,
-        worker: PlaybackWorker,
+        worker: DecodingWorker,
     },
     Paused {
         path: AudioPath,
         position: Duration,
-        worker: PlaybackWorker,
+        worker: DecodingWorker,
     },
     Stopped,
     Invalid,
@@ -688,227 +554,23 @@ enum PreloadState {
     None,
 }
 
-struct PlaybackWorker {
-    actor: ActorHandle<Decode>,
-}
-
-impl PlaybackWorker {
-    fn start(&self) {
-        self.actor.sender().send(Decode::Start).unwrap();
-    }
-
-    fn stop(&self) {
-        self.actor.sender().send(Decode::Stop).unwrap();
-    }
-
-    fn seek(&self, pos: Duration) {
-        self.actor.sender().send(Decode::Seek(pos)).unwrap();
-    }
-
-    fn quit(&self) {
-        let _ = self.actor.sender().send(Decode::Quit);
-    }
-}
-
-impl Drop for PlaybackWorker {
-    fn drop(&mut self) {
-        self.quit();
-    }
-}
-
-const REPORT_POSITION_EACH: Duration = Duration::from_millis(1000);
-
-enum Decode {
-    Start,
-    Stop,
-    Seek(Duration),
-    ReadPacket,
-    FlushPacket,
-    Quit,
-}
-
-enum DecState {
-    Started,
-    Stopped,
-}
-
-struct Decoding {
-    file: AudioFile,
-    source: AudioDecoder,
-    norm_factor: f32,
-    resampler: AudioResampler,
-    samples: SampleBuffer<f32>,
-    events: Sender<PlayerEvent>,
-    this: Sender<Decode>,
-    sink: AudioSink<f32>,
-    volume: VolumeLevel,
-    state: DecState,
-    last_reported_position: Duration,
-}
-
-impl Decoding {
-    fn new(
-        loaded: LoadedPlaybackItem,
-        events: Sender<PlayerEvent>,
-        this: Sender<Decode>,
-        sink: AudioSink<f32>,
-        volume: VolumeLevel,
-    ) -> Self {
-        let LoadedPlaybackItem {
-            file,
-            source,
-            norm_factor,
-        } = loaded;
-        let resampler = AudioResampler::new(
-            // TODO: Make the quality configurable.
-            ResamplingAlgo::SincMediumQuality,
-            ResamplingSpec {
-                channels: source.channels().unwrap().count(),
-                from_rate: source.sample_rate().unwrap() as usize,
-                to_rate: sink.sample_rate() as usize,
-            },
-            1024 * 8,
-        )
-        .unwrap();
-        let samples = {
-            let max_frames = source.max_frames_per_packet().unwrap_or(1024 * 8);
-            let channels = source.channels().unwrap();
-            let rate = source.sample_rate().unwrap();
-            SampleBuffer::new(max_frames, SignalSpec { rate, channels })
-        };
-        Self {
-            file,
-            source,
-            norm_factor,
-            resampler,
-            samples,
-            events,
-            this,
-            sink,
-            volume,
-            state: DecState::Stopped,
-            last_reported_position: Duration::ZERO,
-        }
-    }
-
-    fn frames_to_duration(&self, frames: u64) -> Duration {
-        Duration::from_secs_f64(frames as f64 / self.source.sample_rate().unwrap() as f64)
-    }
-
-    fn report_position(&mut self, position: Duration) {
-        self.events
-            .send(PlayerEvent::Position {
-                path: self.file.path(),
-                position,
-            })
-            .unwrap();
-        self.last_reported_position = position;
-    }
-
-    fn report_current_position(&mut self) {
-        let position = self.frames_to_duration(self.source.current_frame());
-        self.report_position(position);
-    }
-
-    fn report_current_position_if_neeeded(&mut self) {
-        let position = self.frames_to_duration(self.source.current_frame());
-        if position.saturating_sub(self.last_reported_position) > REPORT_POSITION_EACH {
-            self.report_position(position);
-        }
-    }
-
-    fn is_started(&self) -> bool {
-        matches!(self.state, DecState::Started)
-    }
-}
-
-impl Actor for Decoding {
-    type Message = Decode;
-    type Error = Error;
-
-    fn handle(&mut self, msg: Self::Message) -> Result<ActorOp, Self::Error> {
-        match msg {
-            Decode::Start if !self.is_started() => {
-                self.this.send(Decode::ReadPacket)?;
-                self.state = DecState::Started;
-                Ok(ActorOp::Continue)
-            }
-            Decode::Stop if self.is_started() => {
-                self.state = DecState::Stopped;
-                Ok(ActorOp::Continue)
-            }
-            Decode::Seek(pos) => self.handle_seek(pos),
-            Decode::ReadPacket => self.handle_read_packet(),
-            Decode::FlushPacket => self.handle_flush_packet(),
-            Decode::Quit => Ok(ActorOp::Shutdown),
-            _ => Ok(ActorOp::Continue),
-        }
-    }
-}
-
-impl Decoding {
-    fn handle_seek(&mut self, position: Duration) -> Result<ActorOp, Error> {
-        if let Err(err) = self.source.seek(position) {
-            log::error!("failed to seek: {}", err);
-        } else {
-            self.report_current_position();
-        }
-        Ok(ActorOp::Continue)
-    }
-
-    fn handle_read_packet(&mut self) -> Result<ActorOp, Error> {
-        if self.is_started() {
-            if let Some(packet) = self.source.next_packet() {
-                self.samples.copy_interleaved_ref(packet);
-                self.report_current_position_if_neeeded();
-                self.this.send(Decode::FlushPacket)?;
-            } else {
-                self.events.send(PlayerEvent::EndOfTrack)?;
-                return Ok(ActorOp::Shutdown);
-            }
-        }
-        Ok(ActorOp::Continue)
-    }
-
-    fn handle_flush_packet(&mut self) -> Result<ActorOp, Error> {
-        let samples = self.samples.samples();
-
-        // Resample the sample buffer into a rate that the audio output supports.
-        let resampled = self.resampler.resample(samples)?;
-
-        // Apply the global volume level and the normalization factor.
-        let factor = self.norm_factor * self.volume.get();
-        for sample in resampled.iter_mut() {
-            *sample *= factor;
-        }
-
-        // Write into the sink, block until all samples are committed to the ring buffer.
-        self.sink.write_blocking(resampled)?;
-
-        if self.is_started() {
-            self.this.send(Decode::ReadPacket)?;
-        }
-        Ok(ActorOp::Continue)
-    }
-}
-
 #[derive(Clone)]
-struct VolumeLevel {
+pub struct VolumeLevel {
     volume: Arc<AtomicU32>,
 }
 
 impl VolumeLevel {
-    fn new() -> Self {
+    pub fn new() -> Self {
         Self {
             volume: Arc::new(AtomicU32::new(0)),
         }
     }
 
-    fn set(&self, volume: f32) {
+    pub fn set(&self, volume: f32) {
         self.volume.store(volume.to_bits(), Ordering::Relaxed)
     }
 
-    fn get(&self) -> f32 {
+    pub fn get(&self) -> f32 {
         f32::from_bits(self.volume.load(Ordering::Relaxed))
     }
 }
